Waiting answers and players should be inherited from FFAGame class.
