+++ tt_trivia/FFAGame.py
@@ -10,6 +10,14 @@ from Player import Player
 import nextcord
 
 
+# TODO: These should probably be set via .env
+SKIP_THRESHOLD = 2/3
+# # of seconds to wait
+MAX_PLAYERS = 20
+ANSWER_TIME = 20
+WAIT_PLAYERS = 20
+
+
 class GameStatus(enum.Enum):
     STARTING = 0
     GETTING_PLAYERS = 1
@@ -87,7 +95,7 @@ class FFAGame(ABC):
         return self._status
 
     async def _handle_failed_game(self, e: Exception):
-        self._logger.exception(f"Critical failure ancountereD: {e}")
+        self._logger.exception(f"Critical failure encountered: {e}")
         self._flush_tasks()
         self._trivia_bot.cleanup_game(self)
         await self._trivia_bot.say(self._guild_id, "Critical error encountered. Stopping game.")
@@ -104,3 +112,41 @@ class FFAGame(ABC):
             task = self._task_stack.pop()
             if task is not None:
                 task.cancel()
+
+    async def _wait_answers(self):
+        await asyncio.sleep(ANSWER_TIME - 5)
+        start = time.perf_counter()
+        await self._trivia_bot.say(self._guild_id, "5 seconds left!", "countdown5.wav")
+        end = time.perf_counter()
+        # Pad out the full 5 seconds
+        await asyncio.sleep(5 - (end-start))
+        await self._set_status(GameStatus.QUESTION_RESULTS)
+
+    async def _wait_players(self, game_name):
+        self._logger.info("Waiting for players")
+        await self._trivia_bot.say(self.get_guild_id(),
+                                   f"Game starting in {WAIT_PLAYERS} seconds. Type \"play\" to join!\n\n")
+        half_wait = round(WAIT_PLAYERS/2)
+        start = time.perf_counter()
+        await asyncio.sleep(half_wait)
+        await self._trivia_bot.say(self.get_guild_id(),
+                                   f"Game starting in {half_wait} seconds. Type \"play\" to join!\n\n")
+        await asyncio.sleep(half_wait)
+        end = time.perf_counter()
+        self._logger.info(f"Waited {end - start:.4f} seconds. ")
+        # if nobody played, cleanup and exit
+        if self._player_count < 1:
+            await self._trivia_bot.say(self._guild_id, "Nobody wanted to play... sad.")
+            await self._set_status(GameStatus.ENDING)
+            return
+        start_msg = f"\n**Starting {game_name}\t difficulty: {self._questions.get_difficulty()}\t category: {self._questions.get_category()}\n**"
+        start_msg += f"If you wish to skip a question answer \"skip!\". "
+        start_msg += f"The question will be skipped if {SKIP_THRESHOLD:.0%} or more of players vote to skip."
+        start_msg += f" If it's not skipped all skip votes count as an incorrect answer.\n"
+        start_msg += "Game starting momentarily.\nPlayers:"
+        for player in self._players.values():
+            start_msg += f"\n\t- {player.name}"
+        start_msg += "\n\n"
+        await self._trivia_bot.say(self.get_guild_id(), start_msg, "prepare.wav")
+        await asyncio.sleep(5)
+        await self._set_status(GameStatus.ASKING)
